Move list views into more files
