+++ bookwyrm/urls.py
@@ -358,15 +358,15 @@ urlpatterns = [
     re_path(r"^list/saved/?$", views.SavedLists.as_view(), name="saved-lists"),
     re_path(r"^list/(?P<list_id>\d+)(.json)?/?$", views.List.as_view(), name="list"),
     re_path(r"^list/delete/(?P<list_id>\d+)/?$", views.delete_list, name="delete-list"),
-    re_path(r"^list/add-book/?$", views.list.add_book, name="list-add-book"),
+    re_path(r"^list/add-book/?$", views.add_book, name="list-add-book"),
     re_path(
         r"^list/(?P<list_id>\d+)/remove/?$",
-        views.list.remove_book,
+        views.remove_book,
         name="list-remove-book",
     ),
     re_path(
         r"^list-item/(?P<list_item_id>\d+)/set-position$",
-        views.list.set_book_position,
+        views.set_book_position,
         name="list-set-book-position",
     ),
     re_path(
