+++ bookwyrm/views/__init__.py
@@ -61,6 +61,20 @@ from .imports.manually_review import (
     delete_import_item,
 )
 
+# lists
+from .list.curate import Curate
+from .list.lists import Lists, SavedLists, UserLists
+from .list.list import (
+    List,
+    save_list,
+    unsave_list,
+    delete_list,
+    unsafe_embed_list,
+    add_book,
+    remove_book,
+    set_book_position,
+)
+
 # misc views
 from .author import Author, EditAuthor, update_author_from_remote
 from .directory import Directory
@@ -90,8 +104,6 @@ from .group import (
 from .inbox import Inbox
 from .interaction import Favorite, Unfavorite, Boost, Unboost
 from .isbn import Isbn
-from .list import Lists, SavedLists, List, Curate, UserLists
-from .list import save_list, unsave_list, delete_list, unsafe_embed_list
 from .notifications import Notifications
 from .outbox import Outbox
 from .reading import ReadThrough, delete_readthrough, delete_progressupdate
