+++ bookwyrm/views/list/list.py
@@ -18,81 +18,11 @@ from django.views.decorators.clickjacking import xframe_options_exempt
 
 from bookwyrm import book_search, forms, models
 from bookwyrm.activitypub import ActivitypubResponse
-from bookwyrm.lists_stream import ListsStream
 from bookwyrm.settings import PAGE_LENGTH
-from .helpers import is_api_request
-from .helpers import get_user_from_username
+from bookwyrm.views.helpers import is_api_request
 
 
 # pylint: disable=no-self-use
-class Lists(View):
-    """book list page"""
-
-    def get(self, request):
-        """display a book list"""
-        lists = ListsStream().get_list_stream(request.user)
-        paginated = Paginator(lists, 12)
-        data = {
-            "lists": paginated.get_page(request.GET.get("page")),
-            "list_form": forms.ListForm(),
-            "path": "/list",
-        }
-        return TemplateResponse(request, "lists/lists.html", data)
-
-    @method_decorator(login_required, name="dispatch")
-    # pylint: disable=unused-argument
-    def post(self, request):
-        """create a book_list"""
-        form = forms.ListForm(request.POST)
-        if not form.is_valid():
-            return redirect("lists")
-        book_list = form.save()
-        # list should not have a group if it is not group curated
-        if not book_list.curation == "group":
-            book_list.group = None
-            book_list.save(broadcast=False)
-
-        return redirect(book_list.local_path)
-
-
-@method_decorator(login_required, name="dispatch")
-class SavedLists(View):
-    """saved book list page"""
-
-    def get(self, request):
-        """display book lists"""
-        # hide lists with no approved books
-        lists = request.user.saved_lists.order_by("-updated_date")
-
-        paginated = Paginator(lists, 12)
-        data = {
-            "lists": paginated.get_page(request.GET.get("page")),
-            "list_form": forms.ListForm(),
-            "path": "/list",
-        }
-        return TemplateResponse(request, "lists/lists.html", data)
-
-
-@method_decorator(login_required, name="dispatch")
-class UserLists(View):
-    """a user's book list page"""
-
-    def get(self, request, username):
-        """display a book list"""
-        user = get_user_from_username(request.user, username)
-        lists = models.List.privacy_filter(request.user).filter(user=user)
-        paginated = Paginator(lists, 12)
-
-        data = {
-            "user": user,
-            "is_self": request.user.id == user.id,
-            "lists": paginated.get_page(request.GET.get("page")),
-            "list_form": forms.ListForm(),
-            "path": user.local_path + "/lists",
-        }
-        return TemplateResponse(request, "user/lists.html", data)
-
-
 class List(View):
     """book list page"""
 
@@ -252,49 +182,6 @@ class EmbedList(View):
             ),
         }
         return TemplateResponse(request, "lists/embed-list.html", data)
-
-
-@method_decorator(login_required, name="dispatch")
-class Curate(View):
-    """approve or discard list suggestsions"""
-
-    def get(self, request, list_id):
-        """display a pending list"""
-        book_list = get_object_or_404(models.List, id=list_id)
-        book_list.raise_not_editable(request.user)
-
-        data = {
-            "list": book_list,
-            "pending": book_list.listitem_set.filter(approved=False),
-            "list_form": forms.ListForm(instance=book_list),
-        }
-        return TemplateResponse(request, "lists/curate.html", data)
-
-    def post(self, request, list_id):
-        """edit a book_list"""
-        book_list = get_object_or_404(models.List, id=list_id)
-        book_list.raise_not_editable(request.user)
-
-        suggestion = get_object_or_404(models.ListItem, id=request.POST.get("item"))
-        approved = request.POST.get("approved") == "true"
-        if approved:
-            # update the book and set it to be the last in the order of approved books,
-            # before any pending books
-            suggestion.approved = True
-            order_max = (
-                book_list.listitem_set.filter(approved=True).aggregate(Max("order"))[
-                    "order__max"
-                ]
-                or 0
-            ) + 1
-            suggestion.order = order_max
-            increment_order_in_reverse(book_list.id, order_max)
-            suggestion.save()
-        else:
-            deleted_order = suggestion.order
-            suggestion.delete(broadcast=False)
-            normalize_book_list_ordering(book_list.id, start=deleted_order)
-        return redirect("list-curate", book_list.id)
 
 
 @require_POST
